Moved enum  folder outside of model folder
